Removing listGethAccounts from safe personal
