+++ rpc/src/v1/traits/personal.rs
@@ -27,15 +27,11 @@ pub trait Personal: Sized + Send + Sync + 'static {
 	/// Returns accounts information.
 	fn accounts_info(&self, _: Params) -> Result<Value, Error>;
 
-	/// Returns the accounts available for importing from Geth.
-	fn geth_accounts(&self, _: Params) -> Result<Value, Error>;
-
 	/// Should be used to convert object to io delegate.
 	fn to_delegate(self) -> IoDelegate<Self> {
 		let mut delegate = IoDelegate::new(Arc::new(self));
 		delegate.add_method("personal_listAccounts", Personal::accounts);
 		delegate.add_method("personal_accountsInfo", Personal::accounts_info);
-		delegate.add_method("personal_listGethAccounts", Personal::geth_accounts);
 
 		delegate
 	}
@@ -79,6 +75,9 @@ pub trait PersonalAccounts: Sized + Send + Sync + 'static {
 	/// Imports a number of Geth accounts, with the list provided as the argument.
 	fn import_geth_accounts(&self, _: Params) -> Result<Value, Error>;
 
+	/// Returns the accounts available for importing from Geth.
+	fn geth_accounts(&self, _: Params) -> Result<Value, Error>;
+
 	/// Should be used to convert object to io delegate.
 	fn to_delegate(self) -> IoDelegate<Self> {
 		let mut delegate = IoDelegate::new(Arc::new(self));
@@ -92,6 +91,7 @@ pub trait PersonalAccounts: Sized + Send + Sync + 'static {
 		delegate.add_method("personal_setAccountName", PersonalAccounts::set_account_name);
 		delegate.add_method("personal_setAccountMeta", PersonalAccounts::set_account_meta);
 		delegate.add_method("personal_importGethAccounts", PersonalAccounts::import_geth_accounts);
+		delegate.add_method("personal_listGethAccounts", PersonalAccounts::geth_accounts);
 
 		delegate
 	}
